+++ WhatToEat/Areas/Admin/Models/ViewModels/Eatery/OrdersForAdminVM.cs
@@ -9,7 +9,7 @@ namespace WhatToEat.Areas.Admin.Models.ViewModels.Eatery
     {
         public int OrderNumber { get; set; }
         public string Username { get; set; }
-        public decimal Total { get; set; }
+        public int Total { get; set; }
         public Dictionary<string, int> ProductsAndQty { get; set; }
         public DateTime CreatedAt { get; set; }
     }
