+++ WhatToEat/Models/ViewModels/Diary/DiaryVM.cs
@@ -12,7 +12,7 @@ namespace WhatToEat.Models.ViewModels.Diary
         public string ProductName { get; set; }
         public int Quantity { get; set; }
         public int Calorie { get; set; }
-        public decimal Total { get { return Quantity * Calorie; } }
+        public int Total { get { return Quantity * Calorie; } }
         public string Image { get; set; }
     }
 }
